backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "ddgs>=9.10.0",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "litellm>=0.1.0",
    "pydantic-ai>=1.62.0",
    "pydantic-ai-slim[duckduckgo]>=1.62.0",
    "pydantic-settings>=2.13.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "surrealdb>=1.0.8",
    "uvicorn>=0.40.0",
]
